requirements.txt
asgiref==3.4.1
Django==3.2.7
django-rest-auth==0.9.5
djangorestframework==3.12.4
drf-extensions==0.7.1
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
